backend_backup/requirements.txt
cohere==5.5.3
qdrant-client==1.9.1
psycopg2-binary==2.9.9
asyncpg==0.29.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.2.1
python-dotenv==1.0.0
PyMuPDF==1.23.8
beautifulsoup4==4.12.2
pytest==7.4.3